refactor HistoryCaseRevision tests
